Update mbi-service-policies.md
general clean up of markdown and removal of the section "What if my request falls under the second category?"

"Don't worry. If you still have questions or think something might be a defect, we still want you to reach out to us. We'll work together to find a solution that works for everyone."